+++ react/Noop.tsx
@@ -1,0 +1,5 @@
+import { FC } from 'react'
+
+const Noop: FC = () => null
+
+export default Noop
