--- react/ProductAssemblyOptionItemInputValues.tsx
@@ -1,3 +1,0 @@
-import InputValues from './components/ProductAssemblyOptions/ProductAssemblyOptionItemInputValues'
-
-export default InputValues
