+++ react/components/ProductAssemblyOptions/ProductAssemblyOptionInputValues.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render, fireEvent } from '@vtex/test-tools/react'
 import ProductAssemblyOptions from '../../ProductAssemblyOptions'
-import InputValue from './ProductAssemblyOptionItemInputValues'
+import InputValue from './ProductAssemblyOptionInputValues'
 import productAttachment from '../../__mocks__/productAttachment.json'
 import useProduct, { ProductContext } from 'vtex.product-context/useProduct'
 import { useProductDispatch } from 'vtex.product-context/ProductDispatchContext'
@@ -91,3 +91,4 @@ test('should trigger changes to Product Context', () => {
   expect(mockedDispatch.mock.calls[2][0].args.groupInputValues['Front text']).toBe('Foobar')
   expect(mockedDispatch.mock.calls[3][0].args.groupInputValues['Glossy print']).toBe(false)
 })
+
